feat: 템플릿 conroller 분리
