+++ src/main/java/roomescape/controller/ReservationController.java
@@ -4,19 +4,20 @@ import jakarta.validation.Valid;
 import java.net.URI;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.Reservation;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
 import roomescape.service.ReservationService;
 
-@Controller
+@RestController
+@RequestMapping("/reservations")
 public class ReservationController {
 
     private final ReservationService reservationService;
@@ -25,24 +26,15 @@ public class ReservationController {
         this.reservationService = reservationService;
     }
 
-    @GetMapping("/reservation")
-    public String showReservationPage() {
-        return "new-reservation";
-    }
-
-    @GetMapping("/reservations")
-    @ResponseBody
+    @GetMapping
     public List<ReservationResponse> showReservations() {
         return reservationService.findAll().stream()
                 .map(ReservationResponse::from)
                 .toList();
     }
 
-    @PostMapping("/reservations")
-    @ResponseBody
-    public ResponseEntity<ReservationResponse> addReservation(
-            @RequestBody @Valid ReservationRequest request
-    ) {
+    @PostMapping
+    public ResponseEntity<ReservationResponse> addReservation(@RequestBody @Valid ReservationRequest request) {
         Reservation savedReservation = reservationService.create(request);
 
         return ResponseEntity
@@ -50,10 +42,9 @@ public class ReservationController {
                 .body(ReservationResponse.from(savedReservation));
     }
 
-    @DeleteMapping("/reservations/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
         reservationService.delete(id);
-
         return ResponseEntity.noContent().build();
     }
 }
